feat: Destroy session and send back an error if user identity changes
A new feature toggle, default off, causes the session to be deleted when
the user identity on the response does not match the session or request.

There are a small number of requests that cause the user present on the
session at the time of the request to be a different user by the time of
the response. As far as I can tell, these are all cases where a user makes
a login or registration request when a valid session is already open in
their useragent. Previously, there have been some legitimate cases (such
as masquerading) where this could occur, but these all appear to be
accounted for by now.

Because there no longer seems to be any case where this mismatch occurs
and where the response should be allowed through, this commit introduces
a feature toggle `ENFORCE_SAFE_SESSIONS` which will destroy the active
session and overwrite the response.

The plan is to make this behavior available in Maple and permanent in
Nutmeg.